🚧 Broken: Partial i18n revert
