+++ app/dashboard/client/page.tsx
@@ -1,21 +1,16 @@
-import { useTranslations } from "next-intl"
-import { Session, getServerSession } from "next-auth"
+"use client"
 
-import options from "@/auth/options"
+import { useSession } from 'next-auth/react'
 
-export default async function Logic() {
-    return <Page session={await getServerSession(options)} />
-}
-
-function Page({ session }: { session: Session | null }) {
-    const intl = useTranslations()
+export default function Page() {
+    const { data: session, status } = useSession()
 
     return (
         <div className="overflow-hidden rounded-md bg-white shadow">
             <ul role="list" className="divide-y divide-gray-200">
                 <li className="px-6 py-4">
                     <p className="font-semibold">
-                        {intl('page.dashboard.sessionOnServer')}
+                        Session data fetched via a client component
                     </p>
                     <p className="m-2">Hi {session?.user.name}</p>
                 </li>
