+++ app/dashboard/layout.tsx
