+++ app/dashboard/token/page.tsx
@@ -1,5 +1,4 @@
 import Link from "next/link"
-import { useTranslations, useLocale } from "next-intl"
 import { getServerSession } from "next-auth"
 
 import options from "@/auth/options"
@@ -9,34 +8,26 @@ import DeleteToken from "@/components/dashboard/DeleteToken"
 
 import type { ApiToken } from "@prisma/client"
 
-export default async function Logic() {
+export default async function Page() {
     const session = await getServerSession(options)
-
     const tokens: Array<ApiToken> = await prisma.apiToken.findMany({ where: { owner: session?.user.email } })
-
-    return <Page tokens={tokens} />
-}
-
-function Page({ tokens }: { tokens: Array<ApiToken> }) {
-    const intl = useTranslations()
-    const locale = useLocale()
 
     return (
         <div className="divide-y divide-gray-200 rounded-lg bg-white shadow">
             <div className="px-4 py-5 flex gap-y-4 items-left flex-col justify-between md:items-center md:flex-row sm:px-6">
                 <div>
                     <h2 className="text-base font-semibold leading-7 text-gray-900">
-                        {intl('page.dashboard.token.title')}
+                        API Tokens
                     </h2>
                     <p className="mt-1 text-sm leading-6 text-gray-600">
-                        {intl('page.dashboard.token.description')}
+                        Private authorization tokens to request data from our endpoint
                     </p>
                 </div>
 
                 <Link className="rounded-md bg-white px-3.5 py-2.5 text-sm font-semibold text-gray-900 shadow-sm ring-1 ring-inset ring-gray-300 hover:bg-gray-50"
                     href="/dashboard/token/new"
                 >
-                    {intl('page.dashboard.token.generateNewToken')}
+                    Generate new token
                 </Link>
             </div>
             <div className="space-y-6">
@@ -51,7 +42,7 @@ function Page({ tokens }: { tokens: Array<ApiToken> }) {
                                     </div>
                                     <div className="mt-1 flex items-center gap-x-2 text-xs leading-5 text-gray-500">
                                         <p className="whitespace-nowrap">
-                                            {intl('keyword.createdOn')} {new Date(token.createdOn).toLocaleString(locale, { month: "long", day: "numeric", year: "numeric" })}
+                                            Created on {new Date(token.createdOn).toLocaleString('en', { month: "long", day: "numeric", year: "numeric" })}
                                         </p>
                                     </div>
                                 </div>
