+++ app/page.tsx
@@ -1,11 +1,9 @@
 import Link from "next/link"
 import Image from 'next/image'
-import { useTranslations } from 'next-intl'
 
 import aLogo from '@/assets/logo.webp'
 
 export default function Page() {
-    const intl = useTranslations()
 
     return (
         <div className="bg-white">
@@ -22,7 +20,7 @@ export default function Page() {
                     </div>
                     <div className="flex lg:flex-1 lg:justify-end">
                         <Link target="_blank" href="https://github.com/ThijmenGThN/next-leaflet/issues" className="text-sm font-semibold leading-6 text-gray-900">
-                            {intl("page.common.reportIssue")} <span aria-hidden="true">&rarr;</span>
+                            Report an issue <span aria-hidden="true">&rarr;</span>
                         </Link>
                     </div>
                 </nav>
@@ -47,17 +45,17 @@ export default function Page() {
                             next-leaflet
                         </h1>
                         <p className="mt-6 text-lg leading-8 text-gray-600">
-                            {intl('page.common.motto')}
+                            An optimized tech stack for efficiency
                         </p>
                         <div className="mt-10 flex items-center justify-center gap-x-6">
                             <Link
                                 href="/dashboard"
                                 className="rounded-md bg-primary px-3.5 py-2.5 text-sm font-semibold text-white shadow-sm hover:bg-primary-600 focus-visible:outline focus-visible:outline-2 focus-visible:outline-offset-2 focus-visible:outline-primary"
                             >
-                                {intl('keyword.dashboard')}
+                                Dashboard
                             </Link>
                             <Link target="_blank" href="https://github.com/ThijmenGThN/next-leaflet" className="text-sm font-semibold leading-6 text-gray-900">
-                                {intl('keyword.contribute')} <span aria-hidden="true">→</span>
+                                Contribute <span aria-hidden="true">→</span>
                             </Link>
                         </div>
                     </div>
